# apps/reference/specs/openapi/platform-api.yaml
openapi: 3.1.0
info:
  title: Human-Agent API Platform
  version: '2025-01-29'
  description: |
    API platform designed as a first-class experience for both human developers and AI agents.

    **Key Features:**
    - Schema-first design with complete OpenAPI specs
    - Structured error responses with actionable suggestions
    - Agent-aware rate limiting (100 human / 500 agent requests per minute)
    - Header-based versioning (API-Version: YYYY-MM-DD)
    - Dry-run mode for all mutations (?dry_run=true)
    - Automatic agent detection and tracking

    **For AI Agents:**
    - All parameters include descriptions and examples
    - Errors include actionable suggestions for self-correction
    - Rate limit headers enable proactive throttling
    - Retry-After headers prevent retry storms

  contact:
    name: API Platform Team
    email: api-support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development
  - url: https://api.example.com
    description: Production
...
paths:
...
  /api/agents:
    get:
      operationId: listAgents
      summary: List all registered agents
      tags:
        - agents
      description: |
        Returns all registered AI agents. Admin only.
      security:
        - BearerAuth: []

      parameters:
        - $ref: '#/components/parameters/API-Version'
        - $ref: '#/components/parameters/X-Agent-ID'

      responses:
        '200':
          description: List of agents
          headers:
            X-Request-ID:
              $ref: '#/components/headers/X-Request-ID'
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Agent'
                  meta:
                    type: object
                    properties:
                      total:
                        type: integer
                        example: 3

        '400':
          $ref: '#/components/responses/BadRequest'

        '401':
          $ref: '#/components/responses/Unauthorized'

        '403':
          $ref: '#/components/responses/Forbidden'

        '429':
          $ref: '#/components/responses/RateLimitExceeded'
...
  /api/agents/register:
    post:
      operationId: registerAgent
      summary: Register a new AI agent
      tags:
        - agents
      description: |
        Register an AI agent and receive an API key. Admin only.
        Save the API key immediately - it will not be shown again.
      security:
        - BearerAuth: []

      parameters:
        - $ref: '#/components/parameters/API-Version'
        - $ref: '#/components/parameters/X-Agent-ID'
        - $ref: '#/components/parameters/DryRun'

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterAgentRequest'
            examples:
              basic:
                summary: Basic agent registration
                value:
                  name: production-bot
              withRateLimit:
                summary: Agent with custom rate limit
                value:
                  name: high-volume-agent
                  rateLimitOverride: 1000

      responses:
        '201':
          description: Agent registered (save the API key!)
          headers:
            X-Request-ID:
              $ref: '#/components/headers/X-Request-ID'
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/AgentWithKey'
                  warning:
                    type: string
                    example: Save the API key - it will not be shown again!

        '400':
          description: Invalid agent data
          headers:
            X-Request-ID:
              $ref: '#/components/headers/X-Request-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: MISSING_REQUIRED_FIELD
                  message: Agent name is required
                  target: name
                  details:
                    - code: MISSING_NAME
                      message: Agent name must be provided
                      suggestion: Provide a descriptive name for the agent (e.g., "production-bot", "gpt-4-assistant")
                  request_id: req_abc123

        '401':
          $ref: '#/components/responses/Unauthorized'

        '403':
          $ref: '#/components/responses/Forbidden'

        '429':
          $ref: '#/components/responses/RateLimitExceeded'
  /api/agents/{id}:
    get:
      operationId: getAgent
      summary: Get agent details
      tags:
        - agents
      description: |
        Get details of a specific agent by ID. Admin only.
        API key hash is hidden in the response.
      security:
        - BearerAuth: []

      parameters:
        - $ref: '#/components/parameters/API-Version'
        - $ref: '#/components/parameters/X-Agent-ID'
        - name: id
          in: path
          required: true
          description: Agent ID
          schema:
            type: string
            example: agent_abc123

      responses:
        '200':
          description: Agent details
          headers:
            X-Request-ID:
              $ref: '#/components/headers/X-Request-ID'
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Agent'

        '400':
          $ref: '#/components/responses/BadRequest'

        '401':
          $ref: '#/components/responses/Unauthorized'

        '403':
          $ref: '#/components/responses/Forbidden'

        '404':
          description: Agent not found
          headers:
            X-Request-ID:
              $ref: '#/components/headers/X-Request-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: RESOURCE_NOT_FOUND
                  message: Agent not found
                  target: id
                  details:
                    - code: AGENT_NOT_FOUND
                      message: Agent with ID 'agent_xyz' does not exist
                      suggestion: Check the agent ID or use GET /api/agents to list all agents
                  request_id: req_abc123

        '429':
          $ref: '#/components/responses/RateLimitExceeded'

    delete:
      operationId: deleteAgent
      summary: Delete an agent permanently
      tags:
        - agents
      description: |
        Permanently delete an agent. Admin only. This is irreversible.
      security:
        - BearerAuth: []

      parameters:
        - $ref: '#/components/parameters/API-Version'
        - $ref: '#/components/parameters/X-Agent-ID'
        - $ref: '#/components/parameters/DryRun'
        - name: id
          in: path
          required: true
          description: Agent ID
          schema:
            type: string
            example: agent_abc123

      responses:
        '204':
          description: Agent deleted
          headers:
            X-Request-ID:
              $ref: '#/components/headers/X-Request-ID'

        '400':
          $ref: '#/components/responses/BadRequest'

        '401':
          $ref: '#/components/responses/Unauthorized'

        '403':
          $ref: '#/components/responses/Forbidden'

        '404':
          description: Agent not found
          headers:
            X-Request-ID:
              $ref: '#/components/headers/X-Request-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: RESOURCE_NOT_FOUND
                  message: Agent not found
                  target: id
                  details:
                    - code: AGENT_NOT_FOUND
                      message: Agent with ID 'agent_xyz' does not exist
                      suggestion: Check the agent ID or use GET /api/agents to list all agents
                  request_id: req_abc123

        '429':
          $ref: '#/components/responses/RateLimitExceeded'
  /api/agents/{id}/deactivate:
    post:
      operationId: deactivateAgent
      summary: Deactivate an agent
      tags:
        - agents
      description: |
        Revoke an agent's API key without deleting it. Admin only.
        Agent can be reactivated later.
      security:
        - BearerAuth: []

      parameters:
        - $ref: '#/components/parameters/API-Version'
        - $ref: '#/components/parameters/X-Agent-ID'
        - $ref: '#/components/parameters/DryRun'
        - name: id
          in: path
          required: true
          description: Agent ID
          schema:
            type: string
            example: agent_abc123

      responses:
        '200':
          description: Agent deactivated
          headers:
            X-Request-ID:
              $ref: '#/components/headers/X-Request-ID'
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: Agent deactivated successfully
                      agentId:
                        type: string
                        example: agent_abc123

        '400':
          $ref: '#/components/responses/BadRequest'

        '401':
          $ref: '#/components/responses/Unauthorized'

        '403':
          $ref: '#/components/responses/Forbidden'

        '404':
          description: Agent not found
          headers:
            X-Request-ID:
              $ref: '#/components/headers/X-Request-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: RESOURCE_NOT_FOUND
                  message: Agent not found
                  target: id
                  details:
                    - code: AGENT_NOT_FOUND
                      message: Agent with ID 'agent_xyz' does not exist
                      suggestion: Check the agent ID or use GET /api/agents to list all agents
                  request_id: req_abc123

        '429':
          $ref: '#/components/responses/RateLimitExceeded'
  /api/agents/{id}/reactivate:
    post:
      operationId: reactivateAgent
      summary: Reactivate a deactivated agent
      tags:
        - agents
      description: |
        Reactivate a previously deactivated agent. Admin only.
      security:
        - BearerAuth: []

      parameters:
        - $ref: '#/components/parameters/API-Version'
        - $ref: '#/components/parameters/X-Agent-ID'
        - $ref: '#/components/parameters/DryRun'
        - name: id
          in: path
          required: true
          description: Agent ID
          schema:
            type: string
            example: agent_abc123

      responses:
        '200':
          description: Agent reactivated
          headers:
            X-Request-ID:
              $ref: '#/components/headers/X-Request-ID'
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: Agent reactivated successfully
                      agentId:
                        type: string
                        example: agent_abc123

        '400':
          $ref: '#/components/responses/BadRequest'

        '401':
          $ref: '#/components/responses/Unauthorized'

        '403':
          $ref: '#/components/responses/Forbidden'

        '404':
          description: Agent not found
          headers:
            X-Request-ID:
              $ref: '#/components/headers/X-Request-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: RESOURCE_NOT_FOUND
                  message: Agent not found
                  target: id
                  details:
                    - code: AGENT_NOT_FOUND
                      message: Agent with ID 'agent_xyz' does not exist
                      suggestion: Check the agent ID or use GET /api/agents to list all agents
                  request_id: req_abc123

        '429':
          $ref: '#/components/responses/RateLimitExceeded'

  # ==========================================================================
  # Audit
  # ==========================================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /api/auth/login
  parameters:
    API-Version:
      name: API-Version
      in: header
      description: |
        API version in YYYY-MM-DD format.
        Defaults to latest stable version if not specified.
      required: false
      schema:
        type: string
        pattern: '^\d{4}-\d{2}-\d{2}$'
        example: '2025-01-29'
    X-Agent-ID:
      name: X-Agent-ID
      in: header
      description: |
        Optional identifier for the AI agent making the request.
        Used for custom rate limiting and analytics.
      required: false
      schema:
        type: string
        example: 'my-gpt4-agent'
    DryRun:
      name: dry_run
      in: query
      description: |
        Validate the request without executing it.
        Returns validation results and what would happen.
      required: false
      schema:
        type: boolean
        default: false
        example: true
...
  headers:
    X-Request-ID:
      description: Unique request identifier for tracing and support
      schema:
        type: string
        example: req_abc123
...
    X-RateLimit-Limit:
      description: Maximum requests allowed in the current window
      schema:
        type: integer
        example: 500
    X-RateLimit-Remaining:
      description: Requests remaining in the current window
      schema:
        type: integer
        example: 387
...
    Retry-After:
      description: Seconds to wait before retrying (set when rate limited)
      schema:
        type: integer
        example: 42
  schemas:
...
    Agent:
      type: object
      description: Registered AI agent (API key hash hidden)
      properties:
        id:
          type: string
          example: agent_abc123
        name:
          type: string
          example: production-bot
        apiKeyHash:
          type: string
          example: '***hidden***'
        active:
          type: boolean
          example: true
        rateLimitOverride:
          type: integer
          example: 1000
        createdAt:
          type: string
          format: date-time
        lastUsedAt:
          type: string
          format: date-time
    AgentWithKey:
      type: object
      description: Agent with plaintext API key (only returned on registration)
      properties:
        id:
          type: string
          example: agent_abc123
        name:
          type: string
          example: production-bot
        apiKey:
          type: string
          description: Plaintext API key (save immediately, shown only once)
          example: ak_abc123def456...
        active:
          type: boolean
          example: true
        rateLimitOverride:
          type: integer
          example: 1000
        createdAt:
          type: string
          format: date-time
    RegisterAgentRequest:
      type: object
      description: Request body for registering an AI agent
      required:
        - name
      properties:
        name:
          type: string
          description: Descriptive name for the agent
          example: production-bot
        rateLimitOverride:
          type: integer
          minimum: 0
          description: Custom rate limit (requests per minute). Omit for default (500).
          example: 1000

    # Audit schemas
...
    ErrorResponse:
      type: object
      description: Standard error response envelope containing error details
      required:
        - error
      properties:
        error:
          $ref: '#/components/schemas/ApiError'
    ApiError:
      type: object
      description: Detailed error information following RFC 7807 problem details pattern
      required:
        - code
        - message
        - request_id
      properties:
        code:
          type: string
          description: Machine-readable error code
          enum:
            - INVALID_PARAMETER
            - MISSING_REQUIRED_FIELD
            - VALUE_OUT_OF_RANGE
            - INVALID_FORMAT
            - RESOURCE_NOT_FOUND
            - CONFLICT
            - UNAUTHORIZED
            - FORBIDDEN
            - RATE_LIMIT_EXCEEDED
            - INTERNAL_SERVER_ERROR
            - SERVICE_UNAVAILABLE
          example: INVALID_PARAMETER
        message:
          type: string
          description: Human-readable error description
          example: The limit parameter is out of range
        target:
          type: string
          description: The field or parameter that caused the error
          example: limit
        details:
          type: array
          description: Additional error details with actionable suggestions
          items:
            $ref: '#/components/schemas/ErrorDetail'
        doc_url:
          type: string
          format: uri
          description: URL to error documentation
          example: https://docs.example.com/errors/INVALID_PARAMETER
        request_id:
          type: string
          description: Unique request identifier for support
          example: req_abc123
    ErrorDetail:
      type: object
      description: Granular error detail with actionable suggestions for agent self-correction
      required:
        - code
        - message
        - suggestion
      properties:
        code:
          type: string
          description: Specific error detail code
          example: VALUE_OUT_OF_RANGE
        message:
          type: string
          description: Detailed error message
          example: limit must be between 1 and 100
        suggestion:
          type: string
          description: Actionable suggestion for fixing the error (REQUIRED for agent self-correction)
          example: Set limit to a value between 1 and 100, or omit it to use the default of 20
        target:
          type: string
          description: Specific field this detail relates to
          example: limit
  responses:
    BadRequest:
      description: Invalid request parameters
      headers:
        X-Request-ID:
          $ref: '#/components/headers/X-Request-ID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: INVALID_PARAMETER
              message: The limit parameter is out of range
              target: limit
              details:
                - code: VALUE_OUT_OF_RANGE
                  message: limit must be between 1 and 100
                  suggestion: Set limit to a value between 1 and 100, or omit it to use the default of 20
                  target: limit
              request_id: req_abc123
    Unauthorized:
      description: Authentication required
      headers:
        X-Request-ID:
          $ref: '#/components/headers/X-Request-ID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: UNAUTHORIZED
              message: Authentication required
              details:
                - code: MISSING_TOKEN
                  message: No Bearer token provided
                  suggestion: Include Authorization header with a valid Bearer token from /api/auth/login
              request_id: req_abc123
    Forbidden:
      description: Insufficient permissions
      headers:
        X-Request-ID:
          $ref: '#/components/headers/X-Request-ID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: FORBIDDEN
              message: Insufficient permissions
              details:
                - code: PERMISSION_DENIED
                  message: Your role does not have access to this resource
                  suggestion: Contact an admin to request the required permissions
              request_id: req_abc123
...
    RateLimitExceeded:
      description: Rate limit exceeded
      headers:
        X-Request-ID:
          $ref: '#/components/headers/X-Request-ID'
        Retry-After:
          $ref: '#/components/headers/Retry-After'
        X-RateLimit-Limit:
          $ref: '#/components/headers/X-RateLimit-Limit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/X-RateLimit-Remaining'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: RATE_LIMIT_EXCEEDED
              message: Rate limit exceeded
              details:
                - code: TOO_MANY_REQUESTS
                  message: You have exceeded the rate limit of 500 requests per 60 seconds
                  suggestion: Wait 42 seconds before retrying. Implement exponential backoff to avoid retry storms.
                  target: rate_limit
              request_id: req_abc123